project_root/requirements.txt
langchain
langgraph
groq
fastapi
uvicorn
streamlit
pillow
opencv-python
torch
torchvision
matplotlib
python-dotenv